refactor: update checklist and improve ft_memcpy and ft_memmove implementations; add tests for memmove

+++ libft/ft_memcpy.c
@@ -4,12 +4,13 @@ void  *ft_memcpy(void *dest, const void *src, size_t n)
 {
     unsigned char       *d;
     const unsigned char *s;
-    int i;
+    size_t i;
 
     d = dest;
     s = src;
     i = 0;
-    while (i < (int)n)
+
+    while (i < n)
     {
         d[i] = s[i];
         i++;
